+++ SrcEngine/DirectX/SpRootSignature.cpp
@@ -11,7 +11,7 @@ void SpRootSignature::Create()
 	D3D12_ROOT_SIGNATURE_DESC rootSignatureDesc{};
 	rootSignatureDesc.Flags = D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT;
 	rootSignatureDesc.pParameters = params.begin()._Ptr;
-	rootSignatureDesc.NumParameters = params.size();
+	rootSignatureDesc.NumParameters = (UINT)params.size();
 
 	rootSignatureDesc.pStaticSamplers = &samplerDesc;
 	rootSignatureDesc.NumStaticSamplers = 1;
